Initial work to add support for NN potentials to the vdW class -- currently vdW interactions are neglected if both atoms are modeled using a NN potential and the interaction is within the asymmetric unit
