+++ examples/acetamide.NN.properties
@@ -1,0 +1,23 @@
+forcefield AMOEBA-2009
+
+# Correct values:
+a-axis  11.526
+b-axis  11.526
+c-axis  13.589
+
+#a-axis  14.526
+#b-axis  14.526
+#c-axis  18.589
+
+alpha   90.0
+beta    90.0
+gamma  120.0
+spacegroup H3c
+
+vdw-cutoff 12.0
+
+pme-mesh-density 2.0
+pme-order          8
+
+nnterm true
+ani-path ../jtorchani/ANI2xgradients.pt
