+++ examples/acetamide.NN.xyz
@@ -1,0 +1,10 @@
+      9  acetamide
+      1   C    4.68194214    1.30158622    0.51205744   108       2       5       6       7
+      2   C    3.86381708    0.02347683    0.51491470   104       1       8       9
+      3   H    3.85728205   -1.87420050    1.32417836   107       8
+      4   H    5.26662376   -0.94405536    1.69608709   107       8
+      5   H    4.38407657    1.95256909    1.36380295   109       1
+      6   H    5.77402757    1.10458363    0.58411392   109       1
+      7   H    4.50510192    1.86773128   -0.42806617   109       1
+      8   N    4.37032957   -1.01296680    1.24172115   106       2       3       4
+      9   O    2.80539594   -0.06217749   -0.10980979   105       2
