+++ examples/methane.properties
@@ -1,0 +1,5 @@
+forcefield AMOEBA_2004
+
+nnterm true
+ani-path ../jtorchani/ANI2xgradients.pt
+
