+++ examples/methane.xyz
@@ -1,0 +1,6 @@
+      5  Torch ANI Methane Example
+      1   C   -0.06963643   -0.01393854   -0.02839366    22       2       3       4       5
+      2   H   -0.88299629    0.62829890   -0.37399238    23       1
+      3   H   -0.47761179   -0.97151704    0.30337760    23       1
+      4   H    0.44758500    0.46948234    0.80361387    23       1
+      5   H    0.63446049   -0.18195107   -0.84661752    23       1
