+++ modules/potential/src/main/java/ffx/potential/nonbonded/ParticleMeshEwald.java
@@ -37,9 +37,8 @@
 // ******************************************************************************
 package ffx.potential.nonbonded;
 
-import static ffx.numerics.math.ScalarMath.binomial;
-import static ffx.numerics.special.Erf.erfc;
 import static ffx.potential.MolecularAssembly.atomIndexing;
+import static ffx.potential.nonbonded.pme.EwaldParameters.DEFAULT_EWALD_COEFFICIENT;
 import static ffx.potential.parameters.ForceField.ELEC_FORM.PAM;
 import static ffx.potential.parameters.ForceField.toEnumForm;
 import static ffx.potential.parameters.MultipoleType.assignMultipole;
@@ -61,7 +60,6 @@ import static java.lang.String.format;
 import static java.util.Arrays.fill;
 import static java.util.Collections.sort;
 import static org.apache.commons.math3.util.FastMath.max;
-import static org.apache.commons.math3.util.FastMath.min;
 import static org.apache.commons.math3.util.FastMath.pow;
 import static org.apache.commons.math3.util.FastMath.sqrt;
 
@@ -75,11 +73,20 @@ import ffx.numerics.atomic.AtomicDoubleArray3D;
 import ffx.numerics.multipole.MultipoleTensor;
 import ffx.potential.ForceFieldEnergy.Platform;
 import ffx.potential.bonded.Atom;
-import ffx.potential.bonded.Atom.Resolution;
 import ffx.potential.bonded.Bond;
 import ffx.potential.bonded.LambdaInterface;
 import ffx.potential.extended.ExtendedSystem;
+import ffx.potential.nonbonded.pme.AlchemicalParameters;
 import ffx.potential.nonbonded.pme.DirectRegion;
+import ffx.potential.nonbonded.pme.LambdaMode;
+import ffx.potential.nonbonded.pme.PMETimings;
+import ffx.potential.nonbonded.pme.Polarization;
+import ffx.potential.nonbonded.pme.RealSpaceNeighborParameters;
+import ffx.potential.nonbonded.pme.SCFAlgorithm;
+import ffx.potential.nonbonded.pme.SCFPredictor;
+import ffx.potential.nonbonded.pme.SCFPredictorParameters;
+import ffx.potential.nonbonded.pme.ScaleParameters;
+import ffx.potential.nonbonded.pme.EwaldParameters;
 import ffx.potential.nonbonded.pme.ExpandInducedDipolesRegion;
 import ffx.potential.nonbonded.pme.InducedDipoleFieldReduceRegion;
 import ffx.potential.nonbonded.pme.InducedDipoleFieldRegion;
@@ -102,19 +109,14 @@ import ffx.potential.parameters.PolarizeType;
 import ffx.potential.utils.EnergyException;
 import ffx.utilities.Constants;
 import ffx.utilities.FFXKeyword;
-import ffx.utilities.KeywordGroup;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.apache.commons.configuration2.CompositeConfiguration;
-import org.apache.commons.math3.analysis.DifferentiableMultivariateVectorFunction;
-import org.apache.commons.math3.analysis.MultivariateMatrixFunction;
 import org.apache.commons.math3.linear.Array2DRowRealMatrix;
 import org.apache.commons.math3.linear.EigenDecomposition;
-import org.apache.commons.math3.optimization.general.LevenbergMarquardtOptimizer;
 
 /**
  * This Particle Mesh Ewald class implements PME for the AMOEBA polarizable mutlipole force field in
@@ -150,8 +152,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   private static final Logger logger = Logger.getLogger(ParticleMeshEwald.class.getName());
   /** Number of unique tensors for given order. */
   private static final int tensorCount = MultipoleTensor.tensorCount(3);
-  /** The sqrt of PI. */
-  private static final double SQRT_PI = sqrt(Math.PI);
   /**
    * If lambdaTerm is true, some ligand atom interactions with the environment are being turned
    * on/off.
@@ -160,8 +160,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   private final boolean reciprocalSpaceTerm;
   /** Reference to the force field being used. */
   private final ForceField forceField;
-  /** Default cutoff values for PME under periodic boundary conditions. */
-  public static final double DEFAULT_EWALD_CUTOFF = 7.0;
   private static final double DEFAULT_POLAR_EPS = 1.0e-6;
   @FFXKeyword(name = "polar-eps", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0e-6",
       description =
@@ -172,7 +170,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   private final double poleps;
   private final boolean directFallback;
   /** Specify an SCF predictor algorithm. */
-  private final SCFPredictor scfPredictor;
   private final SCFPredictorParameters scfPredictorParameters;
   private final EwaldParameters ewaldParameters;
   private final ScaleParameters scaleParameters;
@@ -324,11 +321,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   private boolean generalizedKirkwoodTerm;
   /** If true, compute coordinate gradient. */
   private boolean gradient = false;
-  /**
-   * If less than 1.0, all atoms with their lambda flags set will have their multipoles and
-   * polarizabilities scaled.
-   */
-  private double lambdaScaleMultipoles = 1.0;
   /** Number of PME multipole interactions. */
   private int interactions;
   /** Number of generalized Kirkwood interactions. */
@@ -447,8 +439,9 @@ public class ParticleMeshEwald implements LambdaInterface {
     printInducedDipoles = properties.getBoolean("pme.printInducedDipoles", false);
     noWindowing = properties.getBoolean("pme.noWindowing", false);
 
-    ewaldParameters = new EwaldParameters(ewaldCutoff);
-    scaleParameters = new ScaleParameters(forceField);
+    double aewald = forceField.getDouble("EWALD_ALPHA", DEFAULT_EWALD_COEFFICIENT);
+    ewaldParameters = new EwaldParameters(ewaldCutoff, aewald);
+    scaleParameters = new ScaleParameters(elecForm, forceField);
     reciprocalSpaceTerm = forceField.getBoolean("RECIPTERM", true);
 
     SCFPredictor scfPredictor;
@@ -459,10 +452,10 @@ public class ParticleMeshEwald implements LambdaInterface {
     } catch (Exception e) {
       scfPredictor = SCFPredictor.NONE;
     }
-    this.scfPredictor = scfPredictor;
-    scfPredictorParameters = new SCFPredictorParameters();
+
+    scfPredictorParameters = new SCFPredictorParameters(scfPredictor, nAtoms);
     if (scfPredictor != SCFPredictor.NONE) {
-      scfPredictorParameters.init();
+      scfPredictorParameters.init(forceField);
     }
 
     String algorithm = forceField.getString("SCF_ALGORITHM", "CG");
@@ -500,7 +493,8 @@ public class ParticleMeshEwald implements LambdaInterface {
 
     pcgSolver = new PCGSolver(maxThreads, poleps, forceField, nAtoms);
 
-    alchemicalParameters = new AlchemicalParameters(forceField, lambdaTerm);
+    alchemicalParameters = new AlchemicalParameters(forceField, lambdaTerm, noWindowing,
+        polarization);
 
     String polar = forceField.getString("POLARIZATION", "MUTUAL");
     if (elecForm == ELEC_FORM.FIXED_CHARGE) {
@@ -535,7 +529,8 @@ public class ParticleMeshEwald implements LambdaInterface {
       sb.append(format("   Polarization:                       %8s\n", polarization.toString()));
       if (polarization == Polarization.MUTUAL) {
         sb.append(format("    SCF Convergence Criteria:         %8.3e\n", poleps));
-        sb.append(format("    SCF Predictor:                     %8s\n", this.scfPredictor));
+        sb.append(format("    SCF Predictor:                     %8s\n",
+            scfPredictorParameters.scfPredictor));
         sb.append(format("    SCF Algorithm:                     %8s\n", scfAlgorithm));
         if (scfAlgorithm == SCFAlgorithm.SOR) {
           sb.append(format("    SOR Parameter:                     %8.3f\n", sorRegion.getSOR()));
@@ -653,38 +648,6 @@ public class ParticleMeshEwald implements LambdaInterface {
     } else {
       generalizedKirkwood = null;
     }
-  }
-
-  /**
-   * ewaldCutoff
-   *
-   * @param coeff a double.
-   * @param maxCutoff a double.
-   * @param eps a double.
-   * @return a double.
-   */
-  public static double ewaldCutoff(double coeff, double maxCutoff, double eps) {
-    // Set the tolerance value; use of 1.0d-8 requires strict convergence of the real Space sum.
-    double ratio = erfc(coeff * maxCutoff) / maxCutoff;
-
-    if (ratio > eps) {
-      return maxCutoff;
-    }
-
-    // Use a binary search to refine the coefficient.
-    double xlo = 0.0;
-    double xhi = maxCutoff;
-    double cutoff = 0.0;
-    for (int j = 0; j < 100; j++) {
-      cutoff = (xlo + xhi) / 2.0;
-      ratio = erfc(coeff * cutoff) / cutoff;
-      if (ratio >= eps) {
-        xlo = cutoff;
-      } else {
-        xhi = cutoff;
-      }
-    }
-    return cutoff;
   }
 
   public void computeInduceDipoleField() {
@@ -821,7 +784,6 @@ public class ParticleMeshEwald implements LambdaInterface {
     initializationRegion.init(
         lambdaTerm,
         extendedSystem,
-        lambdaScaleMultipoles,
         atoms,
         coordinates,
         crystal,
@@ -903,7 +865,7 @@ public class ParticleMeshEwald implements LambdaInterface {
 
     // Log some timings.
     if (logger.isLoggable(Level.FINE)) {
-      pmeTimings.printRealSpaceTimings();
+      pmeTimings.printRealSpaceTimings(maxThreads, realSpaceEnergyRegion);
       if (ewaldParameters.aewald > 0.0 && reciprocalSpaceTerm) {
         reciprocalSpace.printTimings();
       }
@@ -966,22 +928,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   }
 
   /**
-   * getGradient
-   *
-   * @param grad an array of double.
-   */
-  public void getGradients(double[][] grad) {
-    double[] x = grad[0];
-    double[] y = grad[1];
-    double[] z = grad[2];
-    for (int i = 0; i < nAtoms; i++) {
-      x[i] = this.grad.getX(i);
-      y[i] = this.grad.getY(i);
-      z[i] = this.grad.getZ(i);
-    }
-  }
-
-  /**
    * Getter for the field <code>interactions</code>.
    *
    * @return a int.
@@ -1022,10 +968,6 @@ public class ParticleMeshEwald implements LambdaInterface {
           alchemicalParameters.dlPowPol,
           alchemicalParameters.d2lPowPol);
     }
-  }
-
-  public String getName() {
-    return "Cartesian";
   }
 
   public double getPermanentEnergy() {
@@ -1251,34 +1193,6 @@ public class ParticleMeshEwald implements LambdaInterface {
     }
   }
 
-  /**
-   * Pass in atoms that have been assigned electrostatics from a fixed charge force field.
-   *
-   * @param atoms An array of atoms.
-   */
-  public void setFixedCharges(Atom[] atoms) {
-    for (Atom ai : atoms) {
-      if (ai.getResolution() == Resolution.FIXEDCHARGE) {
-        int index = ai.getIndex() - 1;
-        polarizability[index] = 0.0;
-        localMultipole[index][t000] = ai.getMultipoleType().getCharge();
-        localMultipole[index][t100] = 0.0;
-        localMultipole[index][t010] = 0.0;
-        localMultipole[index][t001] = 0.0;
-        localMultipole[index][t200] = 0.0;
-        localMultipole[index][t020] = 0.0;
-        localMultipole[index][t002] = 0.0;
-        localMultipole[index][t110] = 0.0;
-        localMultipole[index][t011] = 0.0;
-        localMultipole[index][t101] = 0.0;
-      }
-    }
-  }
-
-  public void setLambdaMultipoleScale(double multipoleScale) {
-    lambdaScaleMultipoles = multipoleScale;
-  }
-
   private void initAtomArrays() {
     if (localMultipole == null || localMultipole.length < nAtoms) {
       localMultipole = new double[nAtoms][10];
@@ -1324,7 +1238,7 @@ public class ParticleMeshEwald implements LambdaInterface {
       }
       pcgSolver.allocateLists(nSymm, nAtoms);
 
-      if (scfPredictor != SCFPredictor.NONE) {
+      if (scfPredictorParameters.scfPredictor != SCFPredictor.NONE) {
         int predictorOrder = scfPredictorParameters.predictorOrder;
         if (lambdaTerm) {
           scfPredictorParameters.predictorInducedDipole = new double[3][predictorOrder][nAtoms][3];
@@ -1463,7 +1377,8 @@ public class ParticleMeshEwald implements LambdaInterface {
     }
 
     // Force rebuild ligand vapor electrostatics are being computed and vaporCrystal is null.
-    if (alchemicalParameters.doLigandVaporElec && alchemicalParameters.vaporCrystal == null) {
+    if (alchemicalParameters.doLigandVaporElec
+        && alchemicalParameters.vaporCrystal == null) {
       rebuild = true;
     }
 
@@ -1504,7 +1419,8 @@ public class ParticleMeshEwald implements LambdaInterface {
       alchemicalParameters.vaporCrystal.setAperiodic(true);
       NeighborList vacuumNeighborList =
           new NeighborList(
-              null, alchemicalParameters.vaporCrystal, atoms, vacuumOff, 2.0, parallelTeam);
+              null, alchemicalParameters.vaporCrystal, atoms, vacuumOff, 2.0,
+              parallelTeam);
       vacuumNeighborList.setIntermolecular(false, molecule);
 
       alchemicalParameters.vaporLists = new int[1][nAtoms][];
@@ -1515,7 +1431,8 @@ public class ParticleMeshEwald implements LambdaInterface {
         coords[0][i * 3 + 2] = atoms[i].getZ();
       }
       boolean print = logger.isLoggable(Level.FINE);
-      vacuumNeighborList.buildList(coords, alchemicalParameters.vaporLists, isSoft, true, print);
+      vacuumNeighborList.buildList(coords, alchemicalParameters.vaporLists, isSoft,
+          true, print);
       alchemicalParameters.vaporPermanentSchedule = vacuumNeighborList.getPairwiseSchedule();
       alchemicalParameters.vaporEwaldSchedule = alchemicalParameters.vaporPermanentSchedule;
       alchemicalParameters.vacuumRanges = new Range[maxThreads];
@@ -1579,7 +1496,8 @@ public class ParticleMeshEwald implements LambdaInterface {
 
     // Permanent real space is done for the condensed phase. Scale the reciprocal space part.
     alchemicalParameters.doPermanentRealSpace = false;
-    alchemicalParameters.permanentScale = 1.0 - alchemicalParameters.lPowPerm;
+    alchemicalParameters.permanentScale =
+        1.0 - alchemicalParameters.lPowPerm;
     alchemicalParameters.dEdLSign = -1.0;
 
     // If we are past the end of the polarization lambda window, then only the condensed phase is
@@ -1587,7 +1505,8 @@ public class ParticleMeshEwald implements LambdaInterface {
     if (lambda <= alchemicalParameters.polLambdaEnd
         && alchemicalParameters.doNoLigandCondensedSCF) {
       alchemicalParameters.doPolarization = true;
-      alchemicalParameters.polarizationScale = 1.0 - alchemicalParameters.lPowPol;
+      alchemicalParameters.polarizationScale =
+          1.0 - alchemicalParameters.lPowPol;
     } else {
       alchemicalParameters.doPolarization = false;
       alchemicalParameters.polarizationScale = 0.0;
@@ -1627,7 +1546,8 @@ public class ParticleMeshEwald implements LambdaInterface {
     // Scale the permanent vacuum electrostatics. The softcore alpha is not
     // necessary (nothing in vacuum to collide with).
     alchemicalParameters.doPermanentRealSpace = true;
-    alchemicalParameters.permanentScale = 1.0 - alchemicalParameters.lPowPerm;
+    alchemicalParameters.permanentScale =
+        1.0 - alchemicalParameters.lPowPerm;
     alchemicalParameters.dEdLSign = -1.0;
     double lAlphaBack = alchemicalParameters.lAlpha;
     double dlAlphaBack = alchemicalParameters.dlAlpha;
@@ -1640,7 +1560,8 @@ public class ParticleMeshEwald implements LambdaInterface {
     // the condensed phase is necessary.
     if (lambda <= alchemicalParameters.polLambdaEnd) {
       alchemicalParameters.doPolarization = true;
-      alchemicalParameters.polarizationScale = 1.0 - alchemicalParameters.lPowPol;
+      alchemicalParameters.polarizationScale =
+          1.0 - alchemicalParameters.lPowPol;
     } else {
       alchemicalParameters.doPolarization = false;
       alchemicalParameters.polarizationScale = 0.0;
@@ -1682,7 +1603,8 @@ public class ParticleMeshEwald implements LambdaInterface {
       generalizedKirkwood.setLambdaFunction(
           alchemicalParameters.polarizationScale,
           alchemicalParameters.dEdLSign * alchemicalParameters.dlPowPol,
-          alchemicalParameters.dEdLSign * alchemicalParameters.d2lPowPol);
+          alchemicalParameters.dEdLSign
+              * alchemicalParameters.d2lPowPol);
     } else {
       generalizedKirkwoodTerm = false;
     }
@@ -1807,8 +1729,9 @@ public class ParticleMeshEwald implements LambdaInterface {
         }
       }
 
-      if (scfPredictor != SCFPredictor.NONE) {
-        scfPredictorParameters.saveMutualInducedDipoles();
+      if (scfPredictorParameters.scfPredictor != SCFPredictor.NONE) {
+        scfPredictorParameters.saveMutualInducedDipoles(lambdaMode,
+            inducedDipole, inducedDipoleCR, directDipole, directDipoleCR);
       }
 
       if (printInducedDipoles) {
@@ -1880,7 +1803,8 @@ public class ParticleMeshEwald implements LambdaInterface {
         inputDipole, inputDipoleCR, use, molecule,
         ip11, mask12, mask13, mask14, mask15, isSoft, ipdamp, thole, realSpaceNeighborParameters,
         gradient, lambdaTerm, lambdaMode, polarization,
-        ewaldParameters, scaleParameters, alchemicalParameters, pmeTimings.realSpaceEnergyTime,
+        ewaldParameters, scaleParameters, alchemicalParameters,
+        pmeTimings.realSpaceEnergyTime,
         // Output
         grad, torque, lambdaGrad, lambdaTorque, shareddEdLambda, sharedd2EdLambda2);
     realSpaceEnergyRegion.executeWith(parallelTeam);
@@ -1896,7 +1820,8 @@ public class ParticleMeshEwald implements LambdaInterface {
       double eGK = 0.0;
 
       // Create default alchemical parameters.
-      AlchemicalParameters alchemicalParametersGK = new AlchemicalParameters(forceField, false);
+      AlchemicalParameters alchemicalParametersGK = new AlchemicalParameters(
+          forceField, false, noWindowing, polarization);
       // No permanent multipole contribution.
       alchemicalParametersGK.permanentScale = 0.0;
       alchemicalParametersGK.doPermanentRealSpace = false;
@@ -2232,16 +2157,16 @@ public class ParticleMeshEwald implements LambdaInterface {
     }
 
     // Predict the current self-consistent induced dipoles using information from previous steps.
-    if (scfPredictor != SCFPredictor.NONE) {
-      switch (scfPredictor) {
+    if (scfPredictorParameters.scfPredictor != SCFPredictor.NONE) {
+      switch (scfPredictorParameters.scfPredictor) {
         case ASPC:
-          scfPredictorParameters.aspcPredictor();
+          scfPredictorParameters.aspcPredictor(lambdaMode, inducedDipole, inducedDipoleCR);
           break;
         case LS:
-          scfPredictorParameters.leastSquaresPredictor();
+          scfPredictorParameters.leastSquaresPredictor(lambdaMode, inducedDipole, inducedDipoleCR);
           break;
         case POLY:
-          scfPredictorParameters.polynomialPredictor();
+          scfPredictorParameters.polynomialPredictor(lambdaMode, inducedDipole, inducedDipoleCR);
           break;
         default:
           break;
@@ -2547,49 +2472,6 @@ public class ParticleMeshEwald implements LambdaInterface {
   }
 
   /**
-   * A precision of 1.0e-8 results in an Ewald coefficient that ensures continuity in the real space
-   * gradient, but at the cost of increased amplitudes for high frequency reciprocal space structure
-   * factors.
-   */
-  private double ewaldCoefficient(double cutoff, double precision) {
-
-    double eps = 1.0e-8;
-    if (precision < 1.0e-1) {
-      eps = precision;
-    }
-
-    /*
-     * Get an approximate value from cutoff and tolerance.
-     */
-    double ratio = eps + 1.0;
-    double x = 0.5;
-    int i = 0;
-    // Larger values lead to a more "delta-function-like" Gaussian
-    while (ratio >= eps) {
-      i++;
-      x *= 2.0;
-      ratio = erfc(x * cutoff) / cutoff;
-    }
-    /*
-     * Use a binary search to refine the coefficient.
-     */
-    int k = i + 60;
-    double xlo = 0.0;
-    double xhi = x;
-    for (int j = 0; j < k; j++) {
-      x = (xlo + xhi) / 2.0;
-      ratio = erfc(x * cutoff) / cutoff;
-      if (ratio >= eps) {
-        xlo = x;
-      } else {
-        xhi = x;
-      }
-    }
-
-    return x;
-  }
-
-  /**
    * Getter for the field <code>totalMultipoleEnergy</code>.
    *
    * @return a double.
@@ -2708,7 +2590,7 @@ public class ParticleMeshEwald implements LambdaInterface {
       PolarizeType polarizeType = ai.getPolarizeType();
       if (polarizeType != null) {
         if (polarizeType.polarizationGroup != null) {
-          growGroup(group, ai);
+          PolarizeType.growGroup(group, ai);
           sort(group);
           ip11[index] = new int[group.size()];
           int j = 0;
@@ -2723,11 +2605,8 @@ public class ParticleMeshEwald implements LambdaInterface {
           }
         }
       } else {
-        String message =
-            "The polarize keyword was not found for atom "
-                + (index + 1)
-                + " with type "
-                + ai.getType();
+        String message = "The polarize keyword was not found for atom "
+            + (index + 1) + " with type " + ai.getType();
         logger.severe(message);
       }
     }
@@ -2796,36 +2675,6 @@ public class ParticleMeshEwald implements LambdaInterface {
       int j = 0;
       for (int k : list) {
         ip13[i][j++] = k;
-      }
-    }
-  }
-
-  /**
-   * A recursive method that checks all atoms bonded to the seed atom for inclusion in the
-   * polarization group. The method is called on each newly found group member.
-   *
-   * @param group XYZ indeces of current group members.
-   * @param seed The bonds of the seed atom are queried for inclusion in the group.
-   */
-  private void growGroup(List<Integer> group, Atom seed) {
-    List<Bond> bonds = seed.getBonds();
-    for (Bond bond : bonds) {
-      Atom atom = bond.get1_2(seed);
-      int tj = atom.getType();
-      boolean added = false;
-      PolarizeType polarizeType = seed.getPolarizeType();
-      for (int type : polarizeType.polarizationGroup) {
-        if (type == tj) {
-          Integer index = atom.getIndex() - 1;
-          if (!group.contains(index)) {
-            group.add(index);
-            added = true;
-            break;
-          }
-        }
-      }
-      if (added) {
-        growGroup(group, atom);
       }
     }
   }
@@ -3015,1259 +2864,6 @@ public class ParticleMeshEwald implements LambdaInterface {
             eij));
   }
 
-  @FFXKeyword(name = "polarization", clazz = String.class,
-      keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "mutual",
-      description = "[DIRECT / MUTUAL / NONE] "
-          + "Selects between the use of direct and mutual dipole polarization for force fields "
-          + "that incorporate the polarization term. "
-          + "The direct modifier avoids an iterative calculation by using only the permanent "
-          + "electric field in computation of induced dipoles. "
-          + "The mutual option, which is the default in the absence of the polarization property, "
-          + "iterates the induced dipoles to self-consistency."
-          + "The none option turns off polarization and takes precedence over the polarizeterm property.")
-  public enum Polarization {
-    MUTUAL,
-    DIRECT,
-    NONE
-  }
-
-  public enum LambdaMode {
-    OFF,
-    CONDENSED,
-    CONDENSED_NO_LIGAND,
-    VAPOR
-  }
-
-  /**
-   * Describes available SCF algorithms, and whether they are supported by the FFX and/or CUDA
-   * implementations.
-   */
-  public enum SCFAlgorithm {
-    SOR(true, true),
-    CG(true, true),
-    EPT(true, true);
-
-    private final List<Platform> supportedPlatforms;
-
-    SCFAlgorithm(boolean ffx, boolean openMM, Platform... otherPlatforms) {
-      List<Platform> platforms = new ArrayList<>();
-      if (ffx) {
-        platforms.add(Platform.FFX);
-      }
-      if (openMM) {
-        platforms.add(Platform.OMM);
-        platforms.add(Platform.OMM_CUDA);
-        platforms.add(Platform.OMM_REF);
-        // Inapplicable because the OpenCL and optimized C implementations don't have AMOEBA yet.
-        // platforms.add(ForceFieldEnergy.Platform.OMM_OPENCL);
-        // platforms.add(ForceFieldEnergy.Platform.OMM_OPTCPU);
-      }
-      platforms.addAll(Arrays.asList(otherPlatforms));
-      supportedPlatforms = Collections.unmodifiableList(platforms);
-    }
-
-    /**
-     * Returns the list of supported Platforms.
-     *
-     * @return The supported platform List. Unmodifiable.
-     */
-    public List<Platform> getSupportedPlatforms() {
-      return supportedPlatforms;
-    }
-
-    /**
-     * Checks if this platform is supported
-     *
-     * @param platform To check
-     * @return Supported
-     */
-    public boolean isSupported(Platform platform) {
-      return supportedPlatforms.contains(platform);
-    }
-  }
-
-  public enum SCFPredictor {
-    NONE,
-    LS,
-    POLY,
-    ASPC
-  }
-
-  public static class RealSpaceNeighborParameters {
-
-    final int numThreads;
-    /**
-     * Neighbor lists, without atoms beyond the real space cutoff.
-     * [nSymm][nAtoms][nIncludedNeighbors]
-     */
-    public int[][][] realSpaceLists;
-    /** Number of neighboring atoms within the real space cutoff. [nSymm][nAtoms] */
-    public int[][] realSpaceCounts;
-    /** Optimal pairwise ranges. */
-    public Range[] realSpaceRanges;
-    /** Pairwise schedule for load balancing. */
-    public IntegerSchedule realSpaceSchedule;
-
-    public RealSpaceNeighborParameters(int maxThreads) {
-      numThreads = maxThreads;
-      realSpaceRanges = new Range[maxThreads];
-    }
-
-    public void allocate(int nAtoms, int nSymm) {
-      realSpaceSchedule = new PairwiseSchedule(numThreads, nAtoms, realSpaceRanges);
-      realSpaceLists = new int[nSymm][nAtoms][];
-      realSpaceCounts = new int[nSymm][nAtoms];
-    }
-  }
-
-  /** Mutable Particle Mesh Ewald constants. */
-  public class EwaldParameters {
-
-    @FFXKeyword(name = "ewald-alpha", keywordGroup = KeywordGroup.ParticleMeshEwald, defaultValue = "0.545",
-        description =
-            "Sets the value of the Ewald coefficient, which controls the width of the Gaussian screening charges during particle mesh Ewald summation for multipole electrostatics. "
-                + "In the absence of the ewald-alpha keyword, the default value is 0.545, which is appropriate for most applications.")
-    public double aewald;
-    public double aewald3;
-    public double an0;
-    public double an1;
-    public double an2;
-    public double an3;
-    public double an4;
-    public double an5;
-    public double off;
-    public double off2;
-
-    public EwaldParameters(double cutoff) {
-      double aewald = forceField.getDouble("EWALD_ALPHA", 0.545);
-      setEwaldParameters(cutoff, aewald);
-    }
-
-    /**
-     * Determine the real space Ewald parameters and permanent multipole self energy.
-     *
-     * @param off Real space cutoff.
-     * @param aewald Ewald convergence parameter (0.0 turns off reciprocal space).
-     */
-    public void setEwaldParameters(double off, double aewald) {
-      this.off = off;
-      this.aewald = aewald;
-      off2 = off * off;
-      double alsq2 = 2.0 * aewald * aewald;
-      double piEwald = Double.POSITIVE_INFINITY;
-      if (aewald > 0.0) {
-        piEwald = 1.0 / (SQRT_PI * aewald);
-      }
-      aewald3 = 4.0 / 3.0 * pow(aewald, 3.0) / SQRT_PI;
-      if (aewald > 0.0) {
-        an0 = alsq2 * piEwald;
-        an1 = alsq2 * an0;
-        an2 = alsq2 * an1;
-        an3 = alsq2 * an2;
-        an4 = alsq2 * an3;
-        an5 = alsq2 * an4;
-      } else {
-        an0 = 0.0;
-        an1 = 0.0;
-        an2 = 0.0;
-        an3 = 0.0;
-        an4 = 0.0;
-        an5 = 0.0;
-      }
-    }
-  }
-
-  /**
-   * Scale factors and masking rules for electrostatics.
-   */
-  public class ScaleParameters {
-
-    private static final double DEFAULT_MPOLE_12_SCALE = 0.0;
-    private static final double DEFAULT_MPOLE_13_SCALE = 0.0;
-    private static final double DEFAULT_MPOLE_14_SCALE = 1.0;
-    private static final double DEFAULT_MPOLE_15_SCALE = 1.0;
-    private static final double DEFAULT_CHG_12_SCALE = 0.0;
-    private static final double DEFAULT_CHG_13_SCALE = 0.0;
-    private static final double DEFAULT_CHG_14_SCALE = 2.0;
-    private static final double DEFAULT_CHG_15_SCALE = 1.0;
-
-    /** The interaction energy between 1-2 multipoles is scaled by m12scale. */
-    @FFXKeyword(name = "mpole-12-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to permanent atomic multipole "
-                + "electrostatic interactions between 1-2 connected atoms, i.e., atoms that are directly bonded. "
-                + "The default value of 0.0 is used, if the mpole-12-scale property is not given "
-                + "in either the parameter file or the property file.")
-    @FFXKeyword(name = "chg-12-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to charge-charge electrostatic "
-                + "interactions between 1-2 connected atoms, i.e., atoms that are directly bonded. "
-                + "The default value of 0.0 is used, if the chg-12-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double m12scale;
-
-    /** The interaction energy between 1-3 multipoles is scaled by m13scale. */
-    @FFXKeyword(name = "mpole-13-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to permanent atomic multipole "
-                + "electrostatic interactions between 1-3 connected atoms, i.e., atoms separated by two covalent bonds. "
-                + "The default value of 0.0 is used, if the mpole-13-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    @FFXKeyword(name = "chg-13-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to charge-charge electrostatic "
-                + "interactions between 1-3 connected atoms, i.e., atoms separated by two covalent bonds. "
-                + "The default value of 0.0 is used, if the chg-13-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double m13scale;
-
-    /** The interaction energy between 1-4 multipoles is scaled by m14scale. */
-    @FFXKeyword(name = "mpole-14-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to permanent atomic multipole "
-                + "electrostatic interactions between 1-4 connected atoms, i.e., atoms separated by three covalent bonds. "
-                + "The default value of 1.0 is used, if the mpole-14-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    @FFXKeyword(name = "chg-14-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to charge-charge electrostatic "
-                + "interactions between 1-4 connected atoms, i.e., atoms separated by three covalent bonds. "
-                + "The default value of 1.0 is used, if the chg-14-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double m14scale;
-
-    /** The interaction energy between 1-5 multipoles is scaled by m15scale. */
-    @FFXKeyword(name = "mpole-15-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to permanent atomic multipole "
-                + "electrostatic interactions between 1-5 connected atoms, i.e., atoms separated by four covalent bonds. "
-                + "The default value of 1.0 is used, if the mpole-15-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    @FFXKeyword(name = "chg-15-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to charge-charge electrostatic "
-                + "interactions between 1-5 connected atoms, i.e., atoms separated by four covalent bonds. "
-                + "The default value of 1.0 is used, if the chg-15-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double m15scale;
-
-    private static final double DEFAULT_DIRECT_11_SCALE = 0.0;
-    private static final double DEFAULT_DIRECT_12_SCALE = 1.0;
-    private static final double DEFAULT_DIRECT_13_SCALE = 1.0;
-    private static final double DEFAULT_DIRECT_14_SCALE = 1.0;
-
-    /**
-     * DIRECT-11-SCALE factor.
-     */
-    @FFXKeyword(name = "direct-11-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to the permanent (direct) field "
-                + "due to atoms within a polarization group during an induced dipole calculation, "
-                + "i.e., atoms that are in the same polarization group as the atom being polarized. "
-                + "The default value of 0.0 is used, if the direct-11-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double d11scale;
-
-    /**
-     * The DIRECT_12_SCALE factor is assumed to be 1.0. If this assumption is violated by a keyword,
-     * FFX will exit.
-     */
-    @FFXKeyword(name = "direct-12-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to the permanent (direct) field "
-                + "due to atoms in 1-2 polarization groups during an induced dipole calculation, "
-                + "i.e., atoms that are in polarization groups directly connected to the group containing the atom being polarized. "
-                + "The default value of 0.0 is used, if the direct-12-scale keyword is not given"
-                + " in either the parameter file or the property file.")
-    public final double d12scale;
-
-    /**
-     * The DIRECT_13_SCALE factor is assumed to be 1.0. If this assumption is violated by a keyword,
-     * FFX will exit.
-     */
-    @FFXKeyword(name = "direct-13-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to the permanent (direct) field "
-                + "due to atoms in 1-3 polarization groups during an induced dipole calculation, "
-                + "i.e., atoms that are in polarization groups separated by one group from the group containing the atom being polarized. "
-                + "The default value of 0.0 is used, if the direct-13-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double d13scale;
-
-    /**
-     * The DIRECT_14_SCALE factor is assumed to be 1.0. If this assumption is violated by a keyword,
-     * FFX will exit.
-     */
-    @FFXKeyword(name = "direct-14-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to the permanent (direct) field "
-                + "due to atoms in 1-4 polarization groups during an induced dipole calculation, "
-                + "i.e., atoms that are in polarization groups separated by two groups from the group containing the atom being polarized. "
-                + "The default value of 1.0 is used, if the direct-14-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double d14scale;
-
-    private static final double DEFAULT_POLAR_12_SCALE = 0.0;
-    private static final double DEFAULT_POLAR_13_SCALE = 0.0;
-    private static final double DEFAULT_POLAR_14_SCALE = 1.0;
-    private static final double DEFAULT_POLAR_15_SCALE = 1.0;
-
-    /**
-     * The interaction energy between a permanent multipole and polarizable site that are 1-2 is
-     * scaled by p12scale.
-     */
-    @FFXKeyword(name = "polar-12-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-2 connected atoms located in different polarization groups. "
-                + "The default value of 0.0 is used, if the polar-12-scale keyword is not given"
-                + " in either the parameter file or the property file.")
-    public final double p12scale;
-
-    /**
-     * The interaction energy between a permanent multipole and polarizable site that are 1-3 is
-     * scaled by p13scale.
-     */
-    @FFXKeyword(name = "polar-13-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-3 connected atoms located in different polarization groups. "
-                + "The default value of 0.0 is used, if the polar-13-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double p13scale;
-
-    /**
-     * The interaction energy between a permanent multipole and polarizable site that are 1-4 is
-     * scaled by p14scale.
-     */
-    @FFXKeyword(name = "polar-14-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-4 connected atoms located in different polarization groups. "
-                + "The default value of 1.0 is used, if the polar-14-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double p14scale;
-
-    /**
-     * The interaction energy between a permanent multipole and polarizable site that are 1-5 is
-     * scaled by p15scale. Only 1.0 is supported.
-     */
-    @FFXKeyword(name = "polar-15-scale", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-5 connected atoms located in different polarization groups. "
-                + "The default value of 1.0 is used, if the polar-15-scale keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double p15scale;
-
-    private static final double DEFAULT_POLAR_12_INTRA = 0.0;
-
-    private static final double DEFAULT_POLAR_13_INTRA = 0.0;
-
-    private static final double DEFAULT_POLAR_14_INTRA = 0.5;
-
-    private static final double DEFAULT_POLAR_15_INTRA = 1.0;
-
-    /**
-     * An intra-12-scale factor other than 0.0 is not supported and will cause FFX to exit.
-     */
-    @FFXKeyword(name = "polar-12-intra", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-2 connected atoms located in the same polarization group. "
-                + "The default value of 0.0 is used, if the polar-12-intra keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double intra12Scale;
-    /**
-     * An intra-13-scale factor other than 0.0 is not supported and will cause FFX to exit.
-     */
-    @FFXKeyword(name = "polar-13-intra", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-3 connected atoms located in the same polarization group. "
-                + "The default value of 0.0 is used, if the polar-13-intra keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double intra13Scale;
-    /**
-     * Provides a multiplicative scale factor that is applied to polarization interactions between
-     * 1-4 connected atoms located in the same polarization group.
-     */
-    @FFXKeyword(name = "polar-14-intra", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "0.5",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-4 connected atoms located in the same polarization group. "
-                + "The default value of 0.5 is used, if the polar-14-intra keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double intra14Scale;
-    /**
-     * An intra-15-scale factor other than 1.0 is not supported and will cause FFX to exit.
-     */
-    @FFXKeyword(name = "polar-15-intra", keywordGroup = ElectrostaticsFunctionalForm, defaultValue = "1.0",
-        description =
-            "Provides a multiplicative scale factor that is applied to polarization interactions "
-                + "between 1-5 connected atoms located in the same polarization group. "
-                + "The default value of 1.0 is used, if the polar-15-intra keyword is not given "
-                + "in either the parameter file or the property file.")
-    public final double intra15Scale;
-
-    public ScaleParameters(ForceField forceField) {
-      if (elecForm == PAM) {
-        m12scale = forceField.getDouble("MPOLE_12_SCALE", DEFAULT_MPOLE_12_SCALE);
-        m13scale = forceField.getDouble("MPOLE_13_SCALE", DEFAULT_MPOLE_13_SCALE);
-        m14scale = forceField.getDouble("MPOLE_14_SCALE", DEFAULT_MPOLE_14_SCALE);
-        m15scale = forceField.getDouble("MPOLE_15_SCALE", DEFAULT_MPOLE_15_SCALE);
-      } else {
-        double m12 = forceField.getDouble("CHG_12_SCALE", DEFAULT_CHG_12_SCALE);
-        if (m12 > 0.0) {
-          m12 = 1.0 / m12;
-        }
-        m12scale = m12;
-        double m13 = forceField.getDouble("CHG_13_SCALE", DEFAULT_CHG_13_SCALE);
-        if (m13 > 0.0) {
-          m13 = 1.0 / m13;
-        }
-        m13scale = m13;
-        double m14 = forceField.getDouble("CHG_14_SCALE", DEFAULT_CHG_14_SCALE);
-        if (m14 > 0.0) {
-          m14 = 1.0 / m14;
-        }
-        m14scale = m14;
-        double m15 = forceField.getDouble("CHG_15_SCALE", DEFAULT_CHG_15_SCALE);
-        if (m15 > 0.0) {
-          m15 = 1.0 / m15;
-        }
-        m15scale = m15;
-      }
-
-      // Multiplicative scale factors applied to polarization interactions between connected
-      // atoms located in the same polarization group.
-      intra12Scale = forceField.getDouble("POLAR_12_INTRA", DEFAULT_POLAR_12_INTRA);
-      if (intra12Scale != 0.0) {
-        logger.severe(format(" Unsupported polar-12-intra parameter: %8.6f", intra12Scale));
-      }
-      intra13Scale = forceField.getDouble("POLAR_13_INTRA", DEFAULT_POLAR_13_INTRA);
-      if (intra13Scale != 0.0) {
-        logger.severe(format(" Unsupported polar-13-intra parameter: %8.6f", intra13Scale));
-      }
-      intra14Scale = forceField.getDouble("POLAR_14_INTRA", DEFAULT_POLAR_14_INTRA);
-      intra15Scale = forceField.getDouble("POLAR_15_INTRA", DEFAULT_POLAR_15_INTRA);
-      if (intra15Scale != 1.0) {
-        logger.severe(format(" Unsupported polar-15-intra parameter: %8.6f", intra15Scale));
-      }
-
-      // Polarization groups masking rules.
-      d11scale = forceField.getDouble("DIRECT_11_SCALE", DEFAULT_DIRECT_11_SCALE);
-      d12scale = forceField.getDouble("DIRECT_12_SCALE", DEFAULT_DIRECT_12_SCALE);
-      d13scale = forceField.getDouble("DIRECT_13_SCALE", DEFAULT_DIRECT_13_SCALE);
-      d14scale = forceField.getDouble("DIRECT_14_SCALE", DEFAULT_DIRECT_14_SCALE);
-      if (d12scale != 1.0) {
-        logger.severe(format(" Unsupported direct-12-scale parameter: %8.6f", d12scale));
-      }
-      if (d13scale != 1.0) {
-        logger.severe(format(" Unsupported direct-13-scale parameter: %8.6f", d13scale));
-      }
-      if (d14scale != 1.0) {
-        logger.severe(format(" Unsupported direct-14-scale parameter: %8.6f", d14scale));
-      }
-
-      // Polarization energy masking rules.
-      p12scale = forceField.getDouble("POLAR_12_SCALE", DEFAULT_POLAR_12_SCALE);
-      p13scale = forceField.getDouble("POLAR_13_SCALE", DEFAULT_POLAR_13_SCALE);
-      p14scale = forceField.getDouble("POLAR_14_SCALE", DEFAULT_POLAR_14_SCALE);
-      p15scale = forceField.getDouble("POLAR_15_SCALE", DEFAULT_POLAR_15_SCALE);
-      if (p15scale != 1.0) {
-        logger.severe(format(" Unsupported polar-15-scale parameter: %8.6f", p15scale));
-      }
-    }
-  }
-
-  public class AlchemicalParameters {
-
-    /** Constant α in: r' = sqrt(r^2 + α*(1 - L)^2) */
-    public double permLambdaAlpha = 1.0;
-    /** Power on L in front of the pairwise multipole potential. */
-    public double permLambdaExponent = 3.0;
-    /** Begin turning on permanent multipoles at Lambda = 0.4; */
-    public double permLambdaStart = 0.4;
-    /** Finish turning on permanent multipoles at Lambda = 1.0; */
-    public double permLambdaEnd = 1.0;
-    /**
-     * Start turning on polarization later in the Lambda path to prevent SCF convergence problems
-     * when atoms nearly overlap.
-     */
-    public double polLambdaStart = 0.75;
-
-    public double polLambdaEnd = 1.0;
-    /** Power on L in front of the polarization energy. */
-    public double polLambdaExponent = 3.0;
-    /** Intramolecular electrostatics for the ligand in vapor is included by default. */
-    public boolean doLigandVaporElec = true;
-    /** Intramolecular electrostatics for the ligand in done in GK implicit solvent. */
-    public boolean doLigandGKElec = false;
-    /**
-     * Condensed phase SCF without the ligand present is included by default. For DualTopologyEnergy
-     * calculations it can be turned off.
-     */
-    public boolean doNoLigandCondensedSCF = true;
-    /** lAlpha = α*(1 - L)^2 */
-    public double lAlpha = 0.0;
-
-    public double dlAlpha = 0.0;
-    public double d2lAlpha = 0.0;
-    public double dEdLSign = 1.0;
-    /** lPowPerm = L^permanentLambdaExponent */
-    public double lPowPerm = 1.0;
-
-    public double dlPowPerm = 0.0;
-    public double d2lPowPerm = 0.0;
-    public boolean doPermanentRealSpace = true;
-    public double permanentScale = 1.0;
-    /** lPowPol = L^polarizationLambdaExponent */
-    public double lPowPol = 1.0;
-
-    public double dlPowPol = 0.0;
-    public double d2lPowPol = 0.0;
-    public boolean doPolarization = true;
-    /**
-     * When computing the polarization energy at L there are 3 pieces.
-     *
-     * <p>1.) Upol(1) = The polarization energy computed normally (ie. system with ligand).
-     *
-     * <p>2.) Uenv = The polarization energy of the system without the ligand.
-     *
-     * <p>3.) Uligand = The polarization energy of the ligand by itself.
-     *
-     * <p>Upol(L) = L*Upol(1) + (1-L)*(Uenv + Uligand)
-     *
-     * <p>Set polarizationScale to L for part 1. Set polarizationScale to (1-L) for parts 2 and 3.
-     */
-    public double polarizationScale = 1.0;
-    /**
-     * The polarization Lambda value goes from 0.0 .. 1.0 as the global lambda value varies between
-     * polLambdaStart .. polLambadEnd.
-     */
-    private double polLambda = 1.0;
-    /**
-     * The permanent Lambda value goes from 0.0 .. 1.0 as the global lambda value varies between
-     * permLambdaStart .. permLambdaEnd.
-     */
-    private double permLambda = 1.0;
-    /** Boundary conditions for the vapor end of the alchemical path. */
-    private Crystal vaporCrystal = null;
-
-    private int[][][] vaporLists = null;
-    private Range[] vacuumRanges = null;
-    private IntegerSchedule vaporPermanentSchedule = null;
-    private IntegerSchedule vaporEwaldSchedule = null;
-
-    public AlchemicalParameters(ForceField forceField, boolean lambdaTerm) {
-      if (lambdaTerm) {
-        // Values of PERMANENT_LAMBDA_ALPHA below 2 can lead to unstable  trajectories.
-        permLambdaAlpha = forceField.getDouble("PERMANENT_LAMBDA_ALPHA", 2.0);
-        if (permLambdaAlpha < 0.0 || permLambdaAlpha > 3.0) {
-          logger.warning(
-              "Invalid value for permanent-lambda-alpha (<0.0 || >3.0); reverting to 2.0");
-          permLambdaAlpha = 2.0;
-        }
-
-        /*
-         A PERMANENT_LAMBDA_EXPONENT of 2 gives a non-zero d2U/dL2 at the
-         beginning of the permanent schedule. Choosing a power of 3 or
-         greater ensures a smooth dU/dL and d2U/dL2 over the schedule.
-
-         A value of 0.0 is also admissible for when ExtendedSystem is
-         scaling multipoles rather than softcoring them.
-        */
-        permLambdaExponent = forceField.getDouble("PERMANENT_LAMBDA_EXPONENT", 3.0);
-        if (permLambdaExponent < 0.0) {
-          logger.warning("Invalid value for permanent-lambda-exponent (<0.0); reverting to 3.0");
-          permLambdaExponent = 3.0;
-        }
-
-        /*
-         A POLARIZATION_LAMBDA_EXPONENT of 2 gives a non-zero d2U/dL2 at
-         the beginning of the polarization schedule. Choosing a power of 3
-         or greater ensures a smooth dU/dL and d2U/dL2 over the schedule.
-
-         A value of 0.0 is also admissible: when polarization is not being
-         softcored but instead scaled, as by ExtendedSystem.
-        */
-        polLambdaExponent = forceField.getDouble("POLARIZATION_LAMBDA_EXPONENT", 3.0);
-        if (polLambdaExponent < 0.0) {
-          logger.warning("Invalid value for polarization-lambda-exponent (<0.0); reverting to 3.0");
-          polLambdaExponent = 3.0;
-        }
-
-        if (noWindowing) {
-          permLambdaStart = 0.0;
-          polLambdaStart = 0.0;
-          permLambdaEnd = 1.0;
-          polLambdaEnd = 1.0;
-          logger.info(
-              "PME-Cart lambda windowing disabled. Permanent and polarization lambda affect entire [0,1].");
-        } else {
-          // Values of PERMANENT_LAMBDA_START below 0.5 can lead to unstable trajectories.
-          permLambdaStart = forceField.getDouble("PERMANENT_LAMBDA_START", 0.4);
-          if (permLambdaStart < 0.0 || permLambdaStart > 1.0) {
-            logger.warning("Invalid value for perm-lambda-start (<0.0 || >1.0); reverting to 0.4");
-            permLambdaStart = 0.4;
-          }
-
-          // Values of PERMANENT_LAMBDA_END must be greater than permLambdaStart and <= 1.0.
-          permLambdaEnd = forceField.getDouble("PERMANENT_LAMBDA_END", 1.0);
-          if (permLambdaEnd < permLambdaStart || permLambdaEnd > 1.0) {
-            logger.warning("Invalid value for perm-lambda-end (<start || >1.0); reverting to 1.0");
-            permLambdaEnd = 1.0;
-          }
-
-          /*
-           The POLARIZATION_LAMBDA_START defines the point in the lambda
-           schedule when the condensed phase polarization of the ligand
-           begins to be turned on. If the condensed phase polarization
-           is considered near lambda=0, then SCF convergence is slow,
-           even with Thole damping. In addition, 2 (instead of 1)
-           condensed phase SCF calculations are necessary from the
-           beginning of the window to lambda=1.
-          */
-          polLambdaStart = forceField.getDouble("POLARIZATION_LAMBDA_START", 0.75);
-          if (polLambdaStart < 0.0 || polLambdaStart > 1.0) {
-            logger.warning("Invalid value for polarization-lambda-start; reverting to 0.75");
-            polLambdaStart = 0.75;
-          }
-
-          /*
-           The POLARIZATION_LAMBDA_END defines the point in the lambda
-           schedule when the condensed phase polarization of ligand has
-           been completely turned on. Values other than 1.0 have not been tested.
-          */
-          polLambdaEnd = forceField.getDouble("POLARIZATION_LAMBDA_END", 1.0);
-          if (polLambdaEnd < polLambdaStart || polLambdaEnd > 1.0) {
-            logger.warning(
-                "Invalid value for polarization-lambda-end (<start || >1.0); reverting to 1.0");
-            polLambdaEnd = 1.0;
-          }
-        }
-
-        // The LAMBDA_VAPOR_ELEC defines if intramolecular electrostatics of the ligand in vapor
-        // will be considered.
-        doLigandVaporElec = forceField.getBoolean("LIGAND_VAPOR_ELEC", true);
-        doLigandGKElec = forceField.getBoolean("LIGAND_GK_ELEC", false);
-        doNoLigandCondensedSCF = forceField.getBoolean("NO_LIGAND_CONDENSED_SCF", true);
-      }
-    }
-
-    public void printLambdaFactors() {
-      StringBuilder sb = new StringBuilder();
-      sb.append(
-          format(
-              "  (%4s)  mode:%-20s lambda:%.2f  permScale:%.2f  polScale:%.2f  dEdLSign:%s  doPol:%-5b  doPermRS:%-5b",
-              "CART",
-              lambdaMode.toString(),
-              lambda,
-              permanentScale,
-              polarizationScale,
-              format("%+f", dEdLSign).charAt(0),
-              doPolarization,
-              doPermanentRealSpace));
-      sb.append(
-          format(
-              "\n    lAlpha:%.2f,%.2f,%.2f  lPowPerm:%.2f,%.2f,%.2f  lPowPol:%.2f,%.2f,%.2f",
-              lAlpha,
-              dlAlpha,
-              d2lAlpha,
-              lPowPerm,
-              dlPowPerm,
-              d2lPowPerm,
-              lPowPol,
-              dlPowPol,
-              d2lPowPol));
-      sb.append(
-          format(
-              "\n    permExp:%.2f  permAlpha:%.2f  permWindow:%.2f,%.2f  polExp:%.2f  polWindow:%.2f,%.2f",
-              permLambdaExponent,
-              permLambdaAlpha,
-              permLambdaStart,
-              permLambdaEnd,
-              polLambdaExponent,
-              polLambdaStart,
-              polLambdaEnd));
-      logger.info(sb.toString());
-    }
-
-    public String toString() {
-      StringBuilder sb = new StringBuilder("   Alchemical Parameters\n");
-      sb.append(
-          format(
-              "    Permanent Multipole Range:      %5.3f-%5.3f\n", permLambdaStart, permLambdaEnd));
-      sb.append(format("    Permanent Multipole Softcore Alpha:   %5.3f\n", permLambdaAlpha));
-      sb.append(format("    Permanent Multipole Lambda Exponent:  %5.3f\n", permLambdaExponent));
-      if (polarization != Polarization.NONE) {
-        sb.append(format("    Polarization Lambda Exponent:         %5.3f\n", polLambdaExponent));
-        sb.append(
-            format(
-                "    Polarization Range:             %5.3f-%5.3f\n", polLambdaStart, polLambdaEnd));
-        sb.append(format("    Condensed SCF Without Ligand:         %B\n", doNoLigandCondensedSCF));
-      }
-      if (!doLigandGKElec) {
-        sb.append(format("    Vapor Electrostatics:                 %B\n", doLigandVaporElec));
-      } else {
-        sb.append(format("    GK Electrostatics at L=0:             %B\n", doLigandGKElec));
-      }
-      return sb.toString();
-    }
-
-    /*
-     * f = sqrt(r^2 + lAlpha)
-     *
-     * df/dL = -alpha * (1.0 - lambda) / f
-     *
-     * g = 1 / sqrt(r^2 + lAlpha)
-     *
-     * dg/dL = alpha * (1.0 - lambda) / (r^2 + lAlpha)^(3/2)
-     *
-     * define dlAlpha = alpha * 1.0 - lambda)
-     *
-     * then df/dL = -dlAlpha / f and dg/dL = dlAlpha * g^3
-     *
-     * Multipoles are turned on from permLambdaStart .. permLambdaEnd.
-     *
-     * @param lambda
-     */
-    public void update(double lambda) {
-
-      lPowPerm = 1.0;
-      dlPowPerm = 0.0;
-      d2lPowPerm = 0.0;
-      lAlpha = 0.0;
-      dlAlpha = 0.0;
-      d2lAlpha = 0.0;
-      if (lambda < permLambdaStart) {
-        lPowPerm = 0.0;
-      } else if (lambda <= permLambdaEnd) {
-        double permWindow = permLambdaEnd - permLambdaStart;
-        double permLambdaScale = 1.0 / permWindow;
-        permLambda = permLambdaScale * (lambda - permLambdaStart);
-
-        lAlpha = permLambdaAlpha * (1.0 - permLambda) * (1.0 - permLambda);
-        dlAlpha = permLambdaAlpha * (1.0 - permLambda);
-        d2lAlpha = -permLambdaAlpha;
-
-        lPowPerm = pow(permLambda, permLambdaExponent);
-        dlPowPerm = permLambdaExponent * pow(permLambda, permLambdaExponent - 1.0);
-        d2lPowPerm = 0.0;
-        if (permLambdaExponent >= 2.0) {
-          d2lPowPerm =
-              permLambdaExponent
-                  * (permLambdaExponent - 1.0)
-                  * pow(permLambda, permLambdaExponent - 2.0);
-        }
-
-        dlAlpha *= permLambdaScale;
-        d2lAlpha *= (permLambdaScale * permLambdaScale);
-        dlPowPerm *= permLambdaScale;
-        d2lPowPerm *= (permLambdaScale * permLambdaScale);
-      }
-
-      // Polarization is turned on from polarizationLambdaStart .. polarizationLambdaEnd.
-      lPowPol = 1.0;
-      dlPowPol = 0.0;
-      d2lPowPol = 0.0;
-      if (lambda < polLambdaStart) {
-        lPowPol = 0.0;
-      } else if (lambda <= polLambdaEnd) {
-        double polWindow = polLambdaEnd - polLambdaStart;
-        double polLambdaScale = 1.0 / polWindow;
-        polLambda = polLambdaScale * (lambda - polLambdaStart);
-        if (polLambdaExponent > 0.0) {
-          lPowPol = pow(polLambda, polLambdaExponent);
-          if (polLambdaExponent >= 1.0) {
-            dlPowPol = polLambdaExponent * pow(polLambda, polLambdaExponent - 1.0);
-            if (polLambdaExponent >= 2.0) {
-              d2lPowPol =
-                  polLambdaExponent
-                      * (polLambdaExponent - 1.0)
-                      * pow(polLambda, polLambdaExponent - 2.0);
-            }
-          }
-        }
-        // Add the chain rule term due to shrinking the lambda range for the polarization energy.
-        dlPowPol *= polLambdaScale;
-        d2lPowPol *= (polLambdaScale * polLambdaScale);
-      }
-    }
-  }
-
-  public class SCFPredictorParameters {
-
-    /** Induced dipole predictor order. */
-    public int predictorOrder;
-    /** Induced dipole predictor index. */
-    public int predictorStartIndex;
-    /** Induced dipole predictor count. */
-    public int predictorCount;
-    /** Dimensions of [mode][predictorOrder][nAtoms][3] */
-    public double[][][][] predictorInducedDipole;
-    /** Dimensions of [mode][predictorOrder][nAtoms][3] */
-    public double[][][][] predictorInducedDipoleCR;
-
-    public LeastSquaresPredictor leastSquaresPredictor;
-    public LevenbergMarquardtOptimizer leastSquaresOptimizer;
-
-    /** Always-stable predictor-corrector for the mutual induced dipoles. */
-    public void aspcPredictor() {
-
-      if (predictorCount < 6) {
-        return;
-      }
-
-      int mode;
-      switch (lambdaMode) {
-        case OFF:
-        case CONDENSED:
-          mode = 0;
-          break;
-        case CONDENSED_NO_LIGAND:
-          mode = 1;
-          break;
-        case VAPOR:
-          mode = 2;
-          break;
-        default:
-          mode = 0;
-      }
-
-      final double[] aspc = {
-          22.0 / 7.0, -55.0 / 14.0, 55.0 / 21.0, -22.0 / 21.0, 5.0 / 21.0, -1.0 / 42.0
-      };
-
-      // Initialize a pointer into predictor induced dipole array.
-      int index = predictorStartIndex;
-
-      // Expansion loop.
-      for (int k = 0; k < 6; k++) {
-
-        // Set the current predictor coefficient.
-        double c = aspc[k];
-        for (int i = 0; i < nAtoms; i++) {
-          for (int j = 0; j < 3; j++) {
-            inducedDipole[0][i][j] += c * predictorInducedDipole[mode][index][i][j];
-            inducedDipoleCR[0][i][j] += c * predictorInducedDipoleCR[mode][index][i][j];
-          }
-        }
-        index++;
-        if (index >= predictorOrder) {
-          index = 0;
-        }
-      }
-    }
-
-    public void init() {
-      predictorCount = 0;
-      int defaultOrder = 6;
-      predictorOrder = forceField.getInteger("SCF_PREDICTOR_ORDER", defaultOrder);
-      if (scfPredictor == SCFPredictor.LS) {
-        leastSquaresPredictor = new LeastSquaresPredictor();
-        double eps = 1.0e-4;
-        leastSquaresOptimizer =
-            new org.apache.commons.math3.optimization.general.LevenbergMarquardtOptimizer(
-                new org.apache.commons.math3.optimization.SimpleVectorValueChecker(eps, eps));
-      } else if (scfPredictor == SCFPredictor.ASPC) {
-        predictorOrder = 6;
-      }
-      predictorStartIndex = 0;
-    }
-
-    /**
-     * The least-squares predictor with induced dipole information from 8-10 previous steps reduces
-     * the number SCF iterations by ~50%.
-     */
-    public void leastSquaresPredictor() {
-      if (predictorCount < 2) {
-        return;
-      }
-      try {
-        /*
-         The Jacobian and target do not change during the LS optimization,
-         so it's most efficient to update them once before the
-         Least-Squares optimizer starts.
-        */
-        leastSquaresPredictor.updateJacobianAndTarget();
-        int maxEvals = 100;
-        fill(leastSquaresPredictor.initialSolution, 0.0);
-        leastSquaresPredictor.initialSolution[0] = 1.0;
-        org.apache.commons.math3.optimization.PointVectorValuePair optimum =
-            leastSquaresOptimizer.optimize(
-                maxEvals,
-                leastSquaresPredictor,
-                leastSquaresPredictor.calculateTarget(),
-                leastSquaresPredictor.weights,
-                leastSquaresPredictor.initialSolution);
-        double[] optimalValues = optimum.getPoint();
-        if (logger.isLoggable(Level.FINEST)) {
-          logger.finest(format("\n LS RMS:            %10.6f", leastSquaresOptimizer.getRMS()));
-          logger.finest(format(" LS Iterations:     %10d", leastSquaresOptimizer.getEvaluations()));
-          logger.finest(
-              format(" Jacobian Evals:    %10d", leastSquaresOptimizer.getJacobianEvaluations()));
-          logger.finest(format(" Chi Square:        %10.6f", leastSquaresOptimizer.getChiSquare()));
-          logger.finest(" LS Coefficients");
-          for (int i = 0; i < predictorOrder - 1; i++) {
-            logger.finest(format(" %2d  %10.6f", i + 1, optimalValues[i]));
-          }
-        }
-
-        int mode;
-        switch (lambdaMode) {
-          case OFF:
-          case CONDENSED:
-            mode = 0;
-            break;
-          case CONDENSED_NO_LIGAND:
-            mode = 1;
-            break;
-          case VAPOR:
-            mode = 2;
-            break;
-          default:
-            mode = 0;
-        }
-
-        // Initialize a pointer into predictor induced dipole array.
-        int index = predictorStartIndex;
-
-        // Apply the LS coefficients in order to provide an initial guess at the converged induced
-        // dipoles.
-        for (int k = 0; k < predictorOrder - 1; k++) {
-
-          // Set the current coefficient.
-          double c = optimalValues[k];
-          for (int i = 0; i < nAtoms; i++) {
-            for (int j = 0; j < 3; j++) {
-              inducedDipole[0][i][j] += c * predictorInducedDipole[mode][index][i][j];
-              inducedDipoleCR[0][i][j] += c * predictorInducedDipoleCR[mode][index][i][j];
-            }
-          }
-          index++;
-          if (index >= predictorOrder) {
-            index = 0;
-          }
-        }
-      } catch (Exception e) {
-        logger.log(Level.WARNING, " Exception computing predictor coefficients", e);
-      }
-    }
-
-    /** Polynomial predictor for the mutual induced dipoles. */
-    public void polynomialPredictor() {
-
-      if (predictorCount == 0) {
-        return;
-      }
-
-      int mode;
-      switch (lambdaMode) {
-        case OFF:
-        case CONDENSED:
-          mode = 0;
-          break;
-        case CONDENSED_NO_LIGAND:
-          mode = 1;
-          break;
-        case VAPOR:
-          mode = 2;
-          break;
-        default:
-          mode = 0;
-      }
-
-      // Check the number of previous induced dipole vectors available.
-      int n = predictorOrder;
-      if (predictorCount < predictorOrder) {
-        n = predictorCount;
-      }
-
-      // Initialize a pointer into predictor induced dipole array.
-      int index = predictorStartIndex;
-
-      // Initialize the sign of the polynomial expansion.
-      double sign = -1.0;
-
-      // Expansion loop.
-      for (int k = 0; k < n; k++) {
-
-        // Set the current predictor sign and coefficient.
-        sign *= -1.0;
-        double c = sign * binomial(n, k);
-        for (int i = 0; i < nAtoms; i++) {
-          for (int j = 0; j < 3; j++) {
-            inducedDipole[0][i][j] += c * predictorInducedDipole[mode][index][i][j];
-            inducedDipoleCR[0][i][j] += c * predictorInducedDipoleCR[mode][index][i][j];
-          }
-        }
-        index++;
-        if (index >= predictorOrder) {
-          index = 0;
-        }
-      }
-    }
-
-    /** Save the current converged mutual induced dipoles. */
-    public void saveMutualInducedDipoles() {
-
-      int mode;
-      switch (lambdaMode) {
-        case OFF:
-        case CONDENSED:
-          mode = 0;
-          break;
-        case CONDENSED_NO_LIGAND:
-          mode = 1;
-          break;
-        case VAPOR:
-          mode = 2;
-          break;
-        default:
-          mode = 0;
-      }
-
-      // Current induced dipoles are saved before those from the previous step.
-      predictorStartIndex--;
-      if (predictorStartIndex < 0) {
-        predictorStartIndex = predictorOrder - 1;
-      }
-
-      if (predictorCount < predictorOrder) {
-        predictorCount++;
-      }
-
-      for (int i = 0; i < nAtoms; i++) {
-        for (int j = 0; j < 3; j++) {
-          predictorInducedDipole[mode][predictorStartIndex][i][j] =
-              inducedDipole[0][i][j] - directDipole[i][j];
-          predictorInducedDipoleCR[mode][predictorStartIndex][i][j] =
-              inducedDipoleCR[0][i][j] - directDipoleCR[i][j];
-        }
-      }
-    }
-
-    private class LeastSquaresPredictor implements DifferentiableMultivariateVectorFunction {
-
-      double[] weights;
-      double[] target;
-      double[] values;
-      double[][] jacobian;
-      double[] initialSolution;
-      private final MultivariateMatrixFunction multivariateMatrixFunction = this::jacobian;
-
-      LeastSquaresPredictor() {
-        weights = new double[2 * nAtoms * 3];
-        target = new double[2 * nAtoms * 3];
-        values = new double[2 * nAtoms * 3];
-        jacobian = new double[2 * nAtoms * 3][predictorOrder - 1];
-        initialSolution = new double[predictorOrder - 1];
-        fill(weights, 1.0);
-        initialSolution[0] = 1.0;
-      }
-
-      @Override
-      public MultivariateMatrixFunction jacobian() {
-        return multivariateMatrixFunction;
-      }
-
-      @Override
-      public double[] value(double[] variables) {
-        int mode;
-        switch (lambdaMode) {
-          case OFF:
-          case CONDENSED:
-            mode = 0;
-            break;
-          case CONDENSED_NO_LIGAND:
-            mode = 1;
-            break;
-          case VAPOR:
-            mode = 2;
-            break;
-          default:
-            mode = 0;
-        }
-
-        for (int i = 0; i < nAtoms; i++) {
-          int index = 6 * i;
-          values[index] = 0;
-          values[index + 1] = 0;
-          values[index + 2] = 0;
-          values[index + 3] = 0;
-          values[index + 4] = 0;
-          values[index + 5] = 0;
-          int pi = predictorStartIndex + 1;
-          if (pi >= predictorOrder) {
-            pi = 0;
-          }
-          for (int j = 0; j < predictorOrder - 1; j++) {
-            values[index] += variables[j] * predictorInducedDipole[mode][pi][i][0];
-            values[index + 1] += variables[j] * predictorInducedDipole[mode][pi][i][1];
-            values[index + 2] += variables[j] * predictorInducedDipole[mode][pi][i][2];
-            values[index + 3] += variables[j] * predictorInducedDipoleCR[mode][pi][i][0];
-            values[index + 4] += variables[j] * predictorInducedDipoleCR[mode][pi][i][1];
-            values[index + 5] += variables[j] * predictorInducedDipoleCR[mode][pi][i][2];
-            pi++;
-            if (pi >= predictorOrder) {
-              pi = 0;
-            }
-          }
-        }
-        return values;
-      }
-
-      double[] calculateTarget() {
-        return target;
-      }
-
-      void updateJacobianAndTarget() {
-        int mode;
-        switch (lambdaMode) {
-          case OFF:
-          case CONDENSED:
-            mode = 0;
-            break;
-          case CONDENSED_NO_LIGAND:
-            mode = 1;
-            break;
-          case VAPOR:
-            mode = 2;
-            break;
-          default:
-            mode = 0;
-        }
-
-        // Update the target.
-        int index = 0;
-        for (int i = 0; i < nAtoms; i++) {
-          target[index++] = predictorInducedDipole[mode][predictorStartIndex][i][0];
-          target[index++] = predictorInducedDipole[mode][predictorStartIndex][i][1];
-          target[index++] = predictorInducedDipole[mode][predictorStartIndex][i][2];
-          target[index++] = predictorInducedDipoleCR[mode][predictorStartIndex][i][0];
-          target[index++] = predictorInducedDipoleCR[mode][predictorStartIndex][i][1];
-          target[index++] = predictorInducedDipoleCR[mode][predictorStartIndex][i][2];
-        }
-
-        // Update the Jacobian.
-        index = predictorStartIndex + 1;
-        if (index >= predictorOrder) {
-          index = 0;
-        }
-        for (int j = 0; j < predictorOrder - 1; j++) {
-          int ji = 0;
-          for (int i = 0; i < nAtoms; i++) {
-            jacobian[ji++][j] = predictorInducedDipole[mode][index][i][0];
-            jacobian[ji++][j] = predictorInducedDipole[mode][index][i][1];
-            jacobian[ji++][j] = predictorInducedDipole[mode][index][i][2];
-            jacobian[ji++][j] = predictorInducedDipoleCR[mode][index][i][0];
-            jacobian[ji++][j] = predictorInducedDipoleCR[mode][index][i][1];
-            jacobian[ji++][j] = predictorInducedDipoleCR[mode][index][i][2];
-          }
-          index++;
-          if (index >= predictorOrder) {
-            index = 0;
-          }
-        }
-      }
-
-      private double[][] jacobian(double[] variables) {
-        return jacobian;
-      }
-    }
-  }
-
-  public class PMETimings {
-
-    public final long[] realSpacePermTime;
-    public final long[] realSpaceEnergyTime;
-    public final long[] realSpaceSCFTime;
-    /** Timing variables. */
-    private final int numThreads;
-
-    public long realSpacePermTotal, realSpaceEnergyTotal, realSpaceSCFTotal;
-    public long bornRadiiTotal, gkEnergyTotal;
-
-    public PMETimings(int numThreads) {
-      this.numThreads = numThreads;
-      realSpacePermTime = new long[numThreads];
-      realSpaceEnergyTime = new long[numThreads];
-      realSpaceSCFTime = new long[numThreads];
-    }
-
-    public void init() {
-      for (int i = 0; i < numThreads; i++) {
-        realSpacePermTime[i] = 0;
-        realSpaceEnergyTime[i] = 0;
-        realSpaceSCFTime[i] = 0;
-      }
-      realSpacePermTotal = 0;
-      realSpaceEnergyTotal = 0;
-      realSpaceSCFTotal = 0;
-      bornRadiiTotal = 0;
-      gkEnergyTotal = 0;
-    }
-
-    public void printRealSpaceTimings() {
-      double total = (realSpacePermTotal + realSpaceSCFTotal + realSpaceEnergyTotal) * NS2SEC;
-      logger.info(format("\n Real Space: %7.4f (sec)", total));
-      logger.info("           Electric Field");
-      logger.info(" Thread    Direct  SCF     Energy     Counts");
-      long minPerm = Long.MAX_VALUE;
-      long maxPerm = 0;
-      long minSCF = Long.MAX_VALUE;
-      long maxSCF = 0;
-      long minEnergy = Long.MAX_VALUE;
-      long maxEnergy = 0;
-      int minCount = Integer.MAX_VALUE;
-      int maxCount = Integer.MIN_VALUE;
-
-      for (int i = 0; i < maxThreads; i++) {
-        int count = realSpaceEnergyRegion.getCount(i);
-        logger.info(
-            format(
-                "    %3d   %7.4f %7.4f %7.4f %10d",
-                i,
-                realSpacePermTime[i] * NS2SEC,
-                realSpaceSCFTime[i] * NS2SEC,
-                realSpaceEnergyTime[i] * NS2SEC,
-                count));
-        minPerm = min(realSpacePermTime[i], minPerm);
-        maxPerm = max(realSpacePermTime[i], maxPerm);
-        minSCF = min(realSpaceSCFTime[i], minSCF);
-        maxSCF = max(realSpaceSCFTime[i], maxSCF);
-        minEnergy = min(realSpaceEnergyTime[i], minEnergy);
-        maxEnergy = max(realSpaceEnergyTime[i], maxEnergy);
-        minCount = min(count, minCount);
-        maxCount = max(count, maxCount);
-      }
-      logger.info(
-          format(
-              " Min      %7.4f %7.4f %7.4f %10d",
-              minPerm * NS2SEC, minSCF * NS2SEC, minEnergy * NS2SEC, minCount));
-      logger.info(
-          format(
-              " Max      %7.4f %7.4f %7.4f %10d",
-              maxPerm * NS2SEC, maxSCF * NS2SEC, maxEnergy * NS2SEC, maxCount));
-      logger.info(
-          format(
-              " Delta    %7.4f %7.4f %7.4f %10d",
-              (maxPerm - minPerm) * NS2SEC,
-              (maxSCF - minSCF) * NS2SEC,
-              (maxEnergy - minEnergy) * NS2SEC,
-              (maxCount - minCount)));
-      logger.info(
-          format(
-              " Actual   %7.4f %7.4f %7.4f %10d",
-              realSpacePermTotal * NS2SEC,
-              realSpaceSCFTotal * NS2SEC,
-              realSpaceEnergyTotal * NS2SEC,
-              realSpaceEnergyRegion.getInteractions()));
-    }
-  }
-
   /**
    * Flag to indicate use of extended variables that interpolate permanent multipoles and/or atomic
    * polarizability between states (e.g., for protonation and/or tautamers).
@@ -4314,18 +2910,9 @@ public class ParticleMeshEwald implements LambdaInterface {
   }
 
   /**
-   * Flag to indicate if each atom is titrating.
-   */
-  boolean[] isAtomTitrating = null;
-  /**
    * The scalar ESV that is operating on each atom (or 1.0 if the atom is not under ESV control).
    */
   double[] perAtomTitrationESV = null;
-  /**
-   * The index of the ESV that is operating on each atom (or -1 if the atom is not under ESV
-   * control).
-   */
-  Integer[] perAtomESVIndex = null;
 
   /**
    * The partial derivative of each multipole with respect to its titration/tautomer ESV (or 0.0 if
@@ -4445,11 +3032,15 @@ public class ParticleMeshEwald implements LambdaInterface {
       dlfAlpha = alchemicalParameters.dlAlpha;
       d2lfAlpha = alchemicalParameters.d2lAlpha;
       lfPowPerm = alchemicalParameters.permanentScale;
-      dlfPowPerm = alchemicalParameters.dlPowPerm * alchemicalParameters.dEdLSign;
-      d2lfPowPerm = alchemicalParameters.d2lPowPerm * alchemicalParameters.dEdLSign;
+      dlfPowPerm =
+          alchemicalParameters.dlPowPerm * alchemicalParameters.dEdLSign;
+      d2lfPowPerm = alchemicalParameters.d2lPowPerm
+          * alchemicalParameters.dEdLSign;
       lfPowPol = alchemicalParameters.polarizationScale;
-      dlfPowPol = alchemicalParameters.dlPowPol * alchemicalParameters.dEdLSign;
-      d2lfPowPol = alchemicalParameters.d2lPowPol * alchemicalParameters.dEdLSign;
+      dlfPowPol =
+          alchemicalParameters.dlPowPol * alchemicalParameters.dEdLSign;
+      d2lfPowPol =
+          alchemicalParameters.d2lPowPol * alchemicalParameters.dEdLSign;
     }
   }
