+++ modules/potential/src/main/java/ffx/potential/nonbonded/implicit/GKEnergyRegion.java
@@ -69,8 +69,7 @@ import ffx.numerics.multipole.PolarizableMultipole;
 import ffx.numerics.multipole.QIFrame;
 import ffx.potential.bonded.Atom;
 import ffx.potential.nonbonded.GeneralizedKirkwood.NonPolarModel;
-import ffx.potential.nonbonded.ParticleMeshEwald;
-import ffx.potential.nonbonded.ParticleMeshEwald.Polarization;
+import ffx.potential.nonbonded.pme.Polarization;
 import ffx.potential.parameters.ForceField;
 import java.util.List;
 import java.util.logging.Level;
@@ -626,7 +625,7 @@ public class GKEnergyRegion extends ParallelRegion {
         // Compute the permanent GK energy gradient.
         permanentEnergyGradient(i, k);
 
-        if (polarization != ParticleMeshEwald.Polarization.NONE) {
+        if (polarization != Polarization.NONE) {
           // Compute the induced GK energy gradient.
           polarizationEnergyGradient(i, k);
         }
@@ -902,7 +901,7 @@ public class GKEnergyRegion extends ParallelRegion {
 
       // Electrostatic solvation energy of the permanent multipoles in the
       // GK reaction potential of the induced dipoles.
-      if (polarization != ParticleMeshEwald.Polarization.NONE) {
+      if (polarization != Polarization.NONE) {
         double esymi =
             -uxi * (dxk * gux[2] + dyk * guy[2] + dzk * guz[2])
                 - uyi * (dxk * gux[3] + dyk * guy[3] + dzk * guz[3])
@@ -2523,7 +2522,7 @@ public class GKEnergyRegion extends ParallelRegion {
       double dsumdri = dsymdr + 0.5 * (dwipdr + dwkpdr);
       double dbi = 0.5 * rbk * dsumdri;
       double dbk = 0.5 * rbi * dsumdri;
-      if (polarization == ParticleMeshEwald.Polarization.MUTUAL) {
+      if (polarization == Polarization.MUTUAL) {
         dpdx -=
             0.5
                 * (dxi * (pxk * gux[5] + pyk * gux[6] + pzk * gux[7])
