+++ modules/potential/src/main/java/ffx/potential/nonbonded/pme/RealSpaceNeighborParameters.java
@@ -1,0 +1,68 @@
+// ******************************************************************************
+//
+// Title:       Force Field X.
+// Description: Force Field X - Software for Molecular Biophysics.
+// Copyright:   Copyright (c) Michael J. Schnieders 2001-2021.
+//
+// This file is part of Force Field X.
+//
+// Force Field X is free software; you can redistribute it and/or modify it
+// under the terms of the GNU General Public License version 3 as published by
+// the Free Software Foundation.
+//
+// Force Field X is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// Force Field X; if not, write to the Free Software Foundation, Inc., 59 Temple
+// Place, Suite 330, Boston, MA 02111-1307 USA
+//
+// Linking this library statically or dynamically with other modules is making a
+// combined work based on this library. Thus, the terms and conditions of the
+// GNU General Public License cover the whole combination.
+//
+// As a special exception, the copyright holders of this library give you
+// permission to link this library with independent modules to produce an
+// executable, regardless of the license terms of these independent modules, and
+// to copy and distribute the resulting executable under terms of your choice,
+// provided that you also meet, for each linked independent module, the terms
+// and conditions of the license of that module. An independent module is a
+// module which is not derived from or based on this library. If you modify this
+// library, you may extend this exception to your version of the library, but
+// you are not obligated to do so. If you do not wish to do so, delete this
+// exception statement from your version.
+//
+// ******************************************************************************
+package ffx.potential.nonbonded.pme;
+
+import edu.rit.pj.IntegerSchedule;
+import edu.rit.util.Range;
+import ffx.potential.nonbonded.PairwiseSchedule;
+
+public class RealSpaceNeighborParameters {
+
+  public final int numThreads;
+  /**
+   * Neighbor lists, without atoms beyond the real space cutoff. [nSymm][nAtoms][nIncludedNeighbors]
+   */
+  public int[][][] realSpaceLists;
+  /** Number of neighboring atoms within the real space cutoff. [nSymm][nAtoms] */
+  public int[][] realSpaceCounts;
+  /** Optimal pairwise ranges. */
+  public Range[] realSpaceRanges;
+  /** Pairwise schedule for load balancing. */
+  public IntegerSchedule realSpaceSchedule;
+
+  public RealSpaceNeighborParameters(int maxThreads) {
+    numThreads = maxThreads;
+    realSpaceRanges = new Range[maxThreads];
+  }
+
+  public void allocate(int nAtoms, int nSymm) {
+    realSpaceSchedule = new PairwiseSchedule(numThreads, nAtoms, realSpaceRanges);
+    realSpaceLists = new int[nSymm][nAtoms][];
+    realSpaceCounts = new int[nSymm][nAtoms];
+  }
+}
